Week3/requirements.txt
easyocr==1.7.1
matplotlib==3.7.3
numpy==1.24.4
opencv_python==4.8.1.78
opencv_python_headless==4.8.1.78
optuna==3.3.0
pandas==2.0.3
Pillow==10.0.1
Pillow==10.1.0
scipy==1.11.3
tqdm==4.66.1
